split the rest of the models
